dash-app/requirements.txt
numpy
pandas
openpyxl
xlrd
dash
dash-bootstrap-components
dash-bootstrap-templates
dash-bio
gunicorn
mysql-connector-python
bibtexparser
